requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv==1.0.1
pymongo==4.5.0
pydantic==2.7.4
motor==3.3.1
PyPDF2==3.0.0
langchain>=0.3.0
langchain-google-genai==2.1.7
faiss-cpu==1.8.0
streamlit==1.31.0
python-multipart==0.0.9
sentence-transformers==2.7.0